Updated openstack/openstack Project: openstack/nova  05a564dc8071d9cc928129e245cb4f1e57601ebb
support TRACE_FAILONLY env variable

In order to get similar behavior to the way testr out of the box
works, only displaying failing results, all the user to define
TRACE_FAILONLY environment variable (any value will do). After this
point only failures, and things leaking out of stderr will end up on
the screen.

We also need to remove the suppression of the inline fail dumps for
this to have the impact that you'd expect.

Change-Id: I72911adc8a03450522429c52db661ec0a1bc4678