Updated openstack/openstack Project: openstack/python-openstackclient  89a58c65cae12713ce279113404518fa5b55bea9
Add functional tests to osc

Create a script that kicks off function tests that exercise
openstackclient commands against a cloud.

If no keystone/openstack process is detected, a devstack instance
is spun up and the tests are run against that.

There is also a hook added to tox.ini so that we can run these
tests easily from a gate job.

Change-Id: I3cc8b2b800de7ca74af506d2c7e8ee481fa985f0