Updated openstack/openstack Project: openstack/oslo.concurrency  8483811889baa2c9f839c1ccbf74d2c14174bd67
Test with both vanilla and eventlet stdlib

We intend this project's code to function both with and without
eventlet monkey patching, so we should test that way too.  This adds
a separate test run that explicitly enables eventlet and removes
the racy monkey patching that existed before.

I left the eventlet-specific unit test because it's making direct
calls into eventlet, so it is a somewhat different case from
implicitly using monkey patched classes and I'd rather leave a
redundant test than remove it and find out it covered something
the others don't.

Change-Id: Idbe9cdd90e9ce5e38b03ec1c20066928daa9ef00
Closes-Bug: 1367966