Updated openstack/openstack Project: openstack/oslo.i18n  32740b0e2303e2499386fd838aa9e389a3f9151e
Use same indentation in doc/source/usage

The first code example for "Creating an Integration Module"
(http://docs.openstack.org/developer/oslo.i18n/usage.html) is not
rendered correctly, it is because the wrong indentation.

This patch corrects this mistake, and uses a same indentation in cross
of the whole file.

Change-Id: I74189de25d3611ed6a57cf35fee6682b81056071
Closes-Bug: #1374477