Updated openstack/openstack Project: openstack/sahara  9d58c11db84c5edad0526060e9e3b922efb5788b
Make versions list sorted for Vanilla and HDP

Right now the list is not sorted. Probably the versions order in
the list is OS dependent. In my case 2.3.0 came before 1.2.1 in UI,
which is not pretty.

The natural ordering is applied to handle cases like 2.2 < 2.13
(lexicographical order will return that 2.2 > 2.13)

Change-Id: I3a98549ff8a369ea202095e8070f65483e0bd099