Updated openstack/openstack Project: openstack/horizon  05b853dd218de4a19676132ec8de306eca3f0578
show correct link to compute instances in stack resource overview tab

There are two related fixes in the stacks resource overview page.
First, any resource types for which horizon does not have a link for
were getting a URL of "None" in the uuid field, and with this
change the link is suppressed if it is None. The second part involves
adding compute resources to the resource mappings, so that the uuid
field for those links to the instance detail page.

Change-Id: Ib8d51b49ac51092a3d7db6d889f4427167e5adb0
Closes-Bug: 1322162