Updated openstack/openstack Project: openstack/ceilometer  28722719f2f97d2a89b94ff968829e8fd9163021
Include a 'node' key and value in ipmi metadata

The ipmi meters use resource_ids which make it challenging
to select information for all the sensors of a particular type.
This change adds metadata that identifies the physical node on
which a sensor exists.

There are still remaining issues with the naming of ipmi and
sensor-related meters that will need to be addressed separately.

Closes-Bug: 1377157
Change-Id: Ib53755112e4757aa53ab60079f3c555e21a72b63