Updated openstack/openstack Project: openstack/nova  55bd7545042f40430e10d17033c7bc84badc1210
libvirt: convert conn test case to avoid DB usage

The LibvirtConnTestCase uses the database for alot of random
stuff. With a bunch more mocking, it can be converted to use
NoDBTestCase which improves perf from 50s to 9s.

Updates the comment in _get_cpu_numa_config_from_instance to
be explicit about which test cases remain to be converted
after this change.

Closes-bug: #1369516
Change-Id: I1b585cba39913ff4609c6d86dcafc07e6c04e2b8