Updated openstack/openstack Project: openstack-infra/nodepool  00e39427c248725b21e56508cd223ba83734422d
Record interesting info into nova metadata

Sometimes one wants to run a quick command across classes of nodes. This
is made really easy if there are some defining characteristics recorded
in the nova metadata that things like the ansible inventory can pick up
on. Add information to the meta parameter to record that information
in nova.

Change-Id: I3e24f5aa004c5bb8de7ffb757035d64804547f1d