Updated openstack/openstack Project: openstack/nova  a77871700ae77df11b15a38ddf68ab25ba4403ba
Correctly catch InstanceExists in servers create API

When osapi_compute_unique_server_name_scope is set to project
or global an attempt to create a server with a name which
already exists results in an InstanceExists exception being raised.
This was not being handled in the API layer causing a 500
status error code. This patch fixes it so the exception is caught
correctly and a 409 Conflict is returned instead

Change-Id: Ifee3476622b227991a4c2a2780ae7ff23787ccb9
Closes-Bug:1376936