Updated openstack/openstack Project: openstack/python-neutronclient  b5f629f38ab55355aebf2f09a68c5590df3aa498
Don't get keystone session if using noauth

This patch tells neutronclient not to open an authentication session
against keystone if using --os-auth-strategy=noauth.

The intended behaviour in bp keystone-v3-api-support was that
--os-auth-strategy will be completely deprecated and that the absence
of --os-auth-url will signify that noauth is needed. This patch does
not intend to deprecate the strategy option, since that seems to be a
much larger change in the client, so a smaller fix is proposed instead.

This is needed for future Neutron integration tests which will be done
in-tree, since we don't want Neutron to depend on other projects
(including Keystone) during these tests.

Closes-bug: #1375765
Change-Id: I9c61e304f75d3d105e1d66782a7495488eefd5e2