Updated openstack/openstack Project: openstack/horizon  31c00ac5ca683476888329a6943488c01656fba8
Style network address for subnet as required field

It does this by adding the required class to the network address field when the
checkbox is enabled.

I have also made the checkbox behave like a switchable select so when the create
subnet checkbox is checked all fields are visible. They are hidden when create
subnet is unselected.

When the checkbox is disabled the 'subnet details' step in the workflow is
hidden and the workflow can finish there by clicking the create button as all
other information is to do with subnet creation.

Issue highlighted in bug #1371901 was caused by the select.switchable methods in
horizon.forms.js looking at everything with class switchable instead of only
looking at select objects with class switchable.

Change-Id: I92d7b5eee758f0433e3fb61704ddb82f52147ad5
Closes-bug: 1367858
Related-bug: 1371901