Updated openstack/openstack Project: openstack/horizon  9c8314091bc5fa968f38e84b4386ff22d10d67f0
Remove glyphicon-eye-open icon from Description

The glyphicon-eye-open icon is displayed in the Description section
of the 'Create User' and 'Update User' form under the Identify->Users
panel. This is because the no_autocomplete field has been set to
'True' for both the 'Create User' and 'Update User' forms. Enabling
this field causes an input of type="password" and name="fake_password"
to be added inorder to prevent Chrome v34+ from autofilling the form.

Change-Id: Ic628772112aae2d320d8d8342305c33d7ada6822
Closes-Bug: #1371787