Updated openstack/openstack Project: openstack/nova  42245e7fdbb934409c09cb5db9b1b49ed3a032d4
VMware: remove _get_vim() from VMwareAPISession

This method has side-effects which are not needed and brings only
clutter in the driver. The vim object should be accessed as a standard
property of the session object.
This patch is mostly a mechanical replacement and doesn't change any
existing functionality.

This is part of blueprint use-oslo-vmware

Change-Id: I1e86a2e5afc5c0b5921c0248806f434df0cdef2a