Updated openstack/openstack Project: openstack/nova  a4df88a4eee1b380011fd0f3ed94b317c4347da0
Update NoMoreFixedIps message description

The exception NoMoreFixedIps within Nova is a generic exception.
It is used to indicate that the subnets within the network no longer
have any available fixed IPs for use.

The original message was non specific.  This change set provides an
updated message that indicates which network has exhausted its fixed
IP set.

This is important for operators that are deploying instances with
multiple networks.

Change-Id: Ib8b854381b616f83e82bc573f85e109f019897dc
Closes-Bug: #1366982