Updated openstack/openstack Project: openstack/glance  639e7e16880f357345bdfefc2bafe34ea5b66b6d
Fix for Adopt glance.store library in Glance

The store module is removed from glance project and new glance_store
module is created, but the glance project code was not updated
properly for the required changes.

The exception (UnknownScheme) was moved to glance_store, but in
glance image-create api (v1), it still refers to
exception.UnknownScheme in get_store_or_400 method.

Change-Id: I09e78b92d2d44a659412361b4f03b4a988714bf5
Closes-Bug: 1373454