Updated openstack/openstack Project: openstack/nova  0c4a448d81e48bc0b02bf41d3541dd0ce64799bc
VMware: Set vmPathName properly in fake driver

When creating a VM, we pass in config.files.vmPathName as '[ds]'.
However, when this value is subsequently retrieved from a real vSphere
server, it is '[ds] uuid/uuid.vmx'.

This change defaults the value of vmPathName if it's a bare datastore
name to mirror the behaviour of a real vSphere server.

Change-Id: Ide2a4d69fe0ce3214c85b132a8d653f8d5e275f9