Updated openstack/openstack Project: openstack/nova  4c43f4f10754b0e99c0aaf4dfa0c239db9aacfe4
Use image metadata from source volume of a snapshot

When booting an instance with the block device mapping V2 API and
source=snapshot, the image metadata is not retrieved and stored in
instance system_metadata.  Metadata is not exposed on a volume snapshot
so it is grabbed from the volume that the snapshot is from.

Change-Id: I96d6c2198b2f94470ab9dfee5e4bddb13992d8d5
Closes-bug: #1375478