Updated openstack/openstack Project: openstack/cinder  a68f8326a8760c7c309bc7ba631d522060c99f3c
Eqlx fix NoSuchOptError for volume_name_template on clone

The eqlx driver was referencing the volume_name_template
config setting via self.configuration.volume_name_template.
This option is not imported in self.configuration.

The curent preferred method for volume clone is to reference
the passed in name, avoiding the need for the driver to
know what the naming template is all together.

Change-Id: If4963bbe61574c006ad04e5e88d04c7f7bc7cee1
Closes-Bug: 1378369