Updated openstack/openstack Project: openstack/tempest-lib  ecb9244049d6424802990d70a765cc03fd5e260d
Don't edit commit history with migrate from tempest tool

This commit changes the migrate_from_tempest tool to not mess around
with the commit history. The previous approach in the tool isn't
actually something that we can do. The commit history we had when the
repository was created is the canonical past for tempest-lib and we
can't migrate over any other tempest commits directly, this was a
mistake on my part.

Moving forward we'll just have to refer to the commit sha1s from the
tempest repo in the commit message. This script will now just copy the
specified files from tempest and generate a commit message with the
complete list of sha1s from the tempest history.

Change-Id: I7e2b2a0543d113961598b33d23985a1082fbba9a