Updated openstack/openstack Project: openstack/nova  d60497e9f9ce8a6fc54a457eaa2cdab2a817ac90
Cleanup of compute api reboot method

The interaction between the check_instance_state decorator for
the compute api reboot method and the extra checking done within
the method for task states was confusing and in some cases
contradictory.

Also several task states allowed by the decorator were implicitly
rejected because they were not listed in the expected task_states for
the instance save call. This too is fixed.

Change-Id: Iba24f2ed9d2d3161318b4b1fc55c0e97d7aed1c6
Closes-Bug: 1379425