Updated openstack/openstack Project: openstack/python-heatclient  08ddbae505f06d22fd9c941cbcf5db4fce49c445
Fixed null str returned for incorrect parameter

For resources like ResourceGroup, the output parameters are not
validated during stack creation. When a stack has an output parameter
which accesses an incorrect attribute of a resource, the output-show
command returns the string 'null'.
This fix instead returns an error indicating the incorrect attribute
being accessed.

Change-Id: I51eea4738bfefb35644a384ec6d4e9176c8f2e26
Closes-bug: #1368041