Updated openstack/openstack Project: openstack/nova  82271e7ff33858661266518513ae06d6ce1f5f30
Fix libvirt watchdog support

Using the flavor extra_specs property "hw_watchdog_action" was broken.
Scheduling of a new instance always failed with NoValidHostFound error
because of ComputeCapabilitiesFilter, which treated this property as a
host capability to be checked.

Commit f0ff4d51057080e769407e873e5ed212f15b773d caused the problem.

To fix this watchdog_action property is put into 'hw:' scope, so
that it will be ignored by ComputeCapabilitiesFilter in scheduler and
handled in libvirt driver. The doc must be fixed accordingly.

Now driver accepts both 'hw_watchdog_action' and
'hw:watchdog_action', tests were edited for these cases.
Were added TODO items to delete the compat code in the next release.

DocImpact
Closes-Bug: #1367344

Change-Id: Ic5344ec34a130ee5a0ed2c7348af0b9d79e3508e