Updated openstack/openstack Project: openstack/tempest  5c1373da5203fd0823af2d4ee2c8991cd1ea67dd
Fix create_user parameters for v3

When testing against the v3 identity API, the create_user function
definition doesn't sync up to how it's being called. The result is
errors about mismatches in the number of arguments.

This patch updates the function to sync up with how it's being
called, which is also how the definition is written for the v2 API.

Change-Id: I7fd68e774e54933e49b98b9690d803475fcb5d62
Closes-Bug: #1376830