Updated openstack/openstack Project: openstack/nova  e03a0a8c922dfc564055fae9b0606dcb09f909d5
Finish objects conversion in the os-interface API

Most of the API code was using instance objects already except for the
listing, so this updates that.

Updates the test to use a fake instance object rather than a dict to
enforce the change.

Also adds a note on the slightly weird call to common.get_instance
in the show method when the instance isn't used for anything, it's
really a check to make sure the instance exists.

Part of blueprint kilo-objects

Change-Id: Ief17c57afbd81e2b6aef9586b448e8797aaa6544