Updated openstack/openstack Project: openstack/python-openstackclient  79a9331f95979c4ca3c2b8f8bcda7b5d6ef0ddd9
Remove 'links' section from several v3 Identity objects

The links field in the returned objects from the v3 Identity
API aren't really useful, so let's remove them.
Managed to remove most of them from the core API.
I'll likely remove the extension/contribution (oauth/federation)
related ones in another patch.

Also in this patch the code for setting services and projects
was changed. Though not incorrect, it was not needed to copy
the entire returned object, we should just need to pass in
the fields we want to update.

Change-Id: I164ca9ad8b28fa10b291e9115ef40753e387c547