Updated openstack/openstack Project: openstack/cinder  fc57fb68872a92dbcdcf56c20482f20fefe0520f
Remove deprecated use of gettextutils import _

The initial hacking check I wrote allowed users to check in code
using 'from cinder.openstack.common.gettextutils import _' or
'from cinder.i18n import _'.  This was to ease the transition for
code that was in flight with the old import.

Now that we have moved Cinder over to using cinder.i18n and most
of the code that was in flight with the old import has been merged
and/or fixed, I am updating the hacking check to enforce the use of
cinder.i18n and fixing the cases that still had the old import.

Change-Id: Iff6ccac1996c06bbc5defd5139113a2baa0b677c